issue241: eliminate set_runtime(globals()) bridge with direct imports and late-import proxies
- Move all shared workflow/phase helper functions from cli.py to cli_shared.py
- Replace _missing_runtime sentinel pattern in workflow.py and phases_legacy.py
  with late-import proxy functions that delegate through cafe.ui.cli
- Convert set_runtime() to no-op in all command modules
- Add backward-compat re-exports in cli.py for test patch targets
- Add proxy prompt helpers in catalog.py and issues.py for test compatibility
- Fix pre-existing F821 (GitOperations) and F401 (ensure_agent_file_exists) in phase.py
- Clean unused imports in agents.py (json, os, tempfile, prompt_checkbox, etc.)
- All 1880 tests pass; Ruff F821/F401 clean
